requirements.txt
asttokens==2.4.1
backcall==0.2.0
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
flake8==6.1.0
fsspec==2023.12.1
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.12.3
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
jupyter-client==8.6.0
jupyter-core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
pygments==2.17.2
python-dateutil==2.8.2
pytorch-crf==0.7.2
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
sentencepiece==0.1.99
six==1.16.0
stack-data==0.6.3
sympy==1.12
tokenize-rt==5.2.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
triton==2.1.0
typing-extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
